Do not hide reprint of cards when building a deck. Fixes #258 and #203
I don't known if there was a technical goal underneath (didn't find any, nor any problem after testing this change). For a lot of people it seems quite unusual to have to check non-legal (from a format point of view) extensions to see reprinted cards. It should be more user-friendly to simply display them two times in infinite format. The little modal screen displays "reprint of" and "reprinted in" information, so it stays clear for everyone.